news_summarization/.ipynb_checkpoints/llm_interactions-checkpoint.py
# llm_interactions.py

import openai
from cachetools import cached, TTLCache

# Configure a cache with a time-to-live of 1 hour and maximum 100 entries
cache = TTLCache(maxsize=100, ttl=3600)

@cached(cache)
def generate_summary(content: str, limit: int, api_key: str) -> str:
    """
    Generates a summary of the given content with caching and error handling.

    Args:
        content (str): The text to summarize.
        limit (int): Character limit for the summary.
        api_key (str): API key for OpenAI.

    Returns:
        str: Generated summary text or error message.
    """
    openai.api_key = api_key
    try:
        response = openai.ChatCompletion.create(
            model="gpt-3.5-turbo",
            messages=[{"role": "system", "content": f"Summarize in {limit} characters: "}, 
                      {"role": "user", "content": content}]
        )
        return response['choices'][0]['message']['content'][:limit]
    except openai.error.OpenAIError as e:
        return f"Error generating summary: {str(e)}"

# Similar functions for extracting key points, analyzing sentiment, and classifying topics
# can also be cached and error-handled in this way.
